fix(random): align integer dtypes and fix lognormal bug for NumPy parity
- binomial: return int64 instead of double (matches np.random.binomial dtype)
- geometric: return int64 instead of double (matches np.random.geometric dtype)
- poisson: return int64 instead of double (matches np.random.poisson dtype)
- hypergeometric: scalar overload now returns 0-d NDArray instead of raw long
- lognormal: fix bug where mean=0 caused NaN due to incorrect parameter conversion
  - NumPy's lognormal(mean, sigma) is just exp(normal(mean, sigma))
  - Was incorrectly computing transformed parameters leading to log(0)

All integer-returning distributions (poisson, binomial, geometric, hypergeometric,
zipf, logseries, negative_binomial) now return int64 to match NumPy exactly.

Test updates:
- OpenBugs.Random.cs: use GetInt64 instead of GetDouble for integer distributions
- np.random.hypergeometric.Test.cs: update scalar test for NDArray return type

+++ src/NumSharp.Core/RandomSampling/np.random.poisson.cs
@@ -28,9 +28,9 @@ namespace NumSharp
                 throw new ArgumentException("lam must be >= 0", nameof(lam));
 
             if (size.IsScalar || size.IsEmpty)
-                return NDArray.Scalar((double)Knuth(lam));
+                return NDArray.Scalar((long)Knuth(lam));
 
-            var result = new NDArray<double>(size);
+            var result = new NDArray<long>(size);
             unsafe
             {
                 long len = result.size;
